# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HostStatus {
  name: string;
  os: string;
  architecture: string;
  docker: string;
  cpuPercent: number;
  memoryPercent: number;
  memoryUsed: string;
  memoryTotal: string;
}

export interface Overview {
  host: HostStatus;
  running: number;
  total: number;
  healthy: number;
  attention: number;
  lastUpdated: string;
}

export type ServiceCategory = typeof ServiceCategory[keyof typeof ServiceCategory];


export const ServiceCategory = {
  llm: 'llm',
  coding: 'coding',
  automation: 'automation',
  interface: 'interface',
  orchestration: 'orchestration',
  observability: 'observability',
  tool: 'tool',
  memory: 'memory',
  runtime: 'runtime',
} as const;

export type ServiceStatus = typeof ServiceStatus[keyof typeof ServiceStatus];


export const ServiceStatus = {
  running: 'running',
  stopped: 'stopped',
  starting: 'starting',
  warning: 'warning',
} as const;

export type ServiceHealth = typeof ServiceHealth[keyof typeof ServiceHealth];


export const ServiceHealth = {
  healthy: 'healthy',
  degraded: 'degraded',
  offline: 'offline',
  unknown: 'unknown',
} as const;

export interface Service {
  id: string;
  name: string;
  slug: string;
  description: string;
  category: ServiceCategory;
  status: ServiceStatus;
  health: ServiceHealth;
  version: string;
  port: number;
  image: string;
  uptime: string;
  cpuPercent: number;
  memory: string;
  configured: boolean;
  connections: number;
  accent: string;
}

export interface CatalogItem {
  id: string;
  name: string;
  tier: string;
  category: string;
  description: string;
  upstreamUrl: string;
  license: string;
  image: string;
  defaultPort: number;
  installed: boolean;
  accent: string;
}

export interface PortMapping {
  containerPort: number;
  hostPort: number;
  label: string;
}

export interface EnvVarSpec {
  key: string;
  description: string;
  required: boolean;
  defaultValue?: string;
  secret: boolean;
}

export interface ManifestDetail {
  id: string;
  name: string;
  tier: string;
  category: string;
  description: string;
  upstreamUrl: string;
  license: string;
  image: string;
  defaultPort: number;
  ports: PortMapping[];
  env: EnvVarSpec[];
  capabilities: string[];
  compatibleConnections: string[];
  securityNote: string;
}

export interface InstallServiceRequest {
  manifestId: string;
  customPort?: number;
}

export interface RemoveServiceResponse {
  success: boolean;
  serviceId: string;
}

export interface ServiceUpdate {
  /**
     * @minimum 1
     * @maximum 65535
     */
  port?: number;
  image?: string;
  configured?: boolean;
}

export type ServiceActionAction = typeof ServiceActionAction[keyof typeof ServiceActionAction];


export const ServiceActionAction = {
  start: 'start',
  stop: 'stop',
  restart: 'restart',
} as const;

export interface ServiceAction {
  action: ServiceActionAction;
}

export type LogLineStream = typeof LogLineStream[keyof typeof LogLineStream];


export const LogLineStream = {
  system: 'system',
  stdout: 'stdout',
  stderr: 'stderr',
} as const;

export interface LogLine {
  id: string;
  timestamp: string;
  stream: LogLineStream;
  message: string;
}

export type ConnectionStatus = typeof ConnectionStatus[keyof typeof ConnectionStatus];


export const ConnectionStatus = {
  connected: 'connected',
  pending: 'pending',
  'needs-config': 'needs-config',
} as const;

export interface Connection {
  id: string;
  source: string;
  target: string;
  protocol: string;
  status: ConnectionStatus;
  detail: string;
}