# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Get the control plane overview
 */
export const GetOverviewResponse = zod.object({
  "host": zod.object({
  "name": zod.string(),
  "os": zod.string(),
  "architecture": zod.string(),
  "docker": zod.string(),
  "cpuPercent": zod.number(),
  "memoryPercent": zod.number(),
  "memoryUsed": zod.string(),
  "memoryTotal": zod.string()
}),
  "running": zod.number().int(),
  "total": zod.number().int(),
  "healthy": zod.number().int(),
  "attention": zod.number().int(),
  "lastUpdated": zod.string()
})


/**
 * @summary List managed services
 */
export const ListServicesResponseItem = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "slug": zod.string(),
  "description": zod.string(),
  "category": zod.enum(['llm', 'coding', 'automation', 'interface', 'orchestration', 'observability', 'tool', 'memory', 'runtime']),
  "status": zod.enum(['running', 'stopped', 'starting', 'warning']),
  "health": zod.enum(['healthy', 'degraded', 'offline', 'unknown']),
  "version": zod.string(),
  "port": zod.number().int(),
  "image": zod.string(),
  "uptime": zod.string(),
  "cpuPercent": zod.number(),
  "memory": zod.string(),
  "configured": zod.boolean(),
  "connections": zod.number().int(),
  "accent": zod.string()
})
export const ListServicesResponse = zod.array(ListServicesResponseItem)


/**
 * @summary List ecosystem catalog entries
 */
export const ListCatalogResponseItem = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "tier": zod.string(),
  "category": zod.string(),
  "description": zod.string(),
  "upstreamUrl": zod.string(),
  "license": zod.string(),
  "image": zod.string(),
  "defaultPort": zod.number().int(),
  "installed": zod.boolean(),
  "accent": zod.string()
})
export const ListCatalogResponse = zod.array(ListCatalogResponseItem)


/**
 * @summary Get detailed manifest for a catalog item
 */
export const GetCatalogManifestParams = zod.object({
  "manifestId": zod.coerce.string()
})

export const GetCatalogManifestResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "tier": zod.string(),
  "category": zod.string(),
  "description": zod.string(),
  "upstreamUrl": zod.string(),
  "license": zod.string(),
  "image": zod.string(),
  "defaultPort": zod.number().int(),
  "ports": zod.array(zod.object({
  "containerPort": zod.number().int(),
  "hostPort": zod.number().int(),
  "label": zod.string()
})),
  "env": zod.array(zod.object({
  "key": zod.string(),
  "description": zod.string(),
  "required": zod.boolean(),
  "defaultValue": zod.string().optional(),
  "secret": zod.boolean()
})),
  "capabilities": zod.array(zod.string()),
  "compatibleConnections": zod.array(zod.string()),
  "securityNote": zod.string()
})


/**
 * @summary Install a service manifest into the managed fleet
 */
export const InstallServiceBody = zod.object({
  "manifestId": zod.string(),
  "customPort": zod.number().int().optional()
})

export const InstallServiceResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "slug": zod.string(),
  "description": zod.string(),
  "category": zod.enum(['llm', 'coding', 'automation', 'interface', 'orchestration', 'observability', 'tool', 'memory', 'runtime']),
  "status": zod.enum(['running', 'stopped', 'starting', 'warning']),
  "health": zod.enum(['healthy', 'degraded', 'offline', 'unknown']),
  "version": zod.string(),
  "port": zod.number().int(),
  "image": zod.string(),
  "uptime": zod.string(),
  "cpuPercent": zod.number(),
  "memory": zod.string(),
  "configured": zod.boolean(),
  "connections": zod.number().int(),
  "accent": zod.string()
})


/**
 * @summary Get one managed service
 */
export const GetServiceParams = zod.object({
  "serviceId": zod.coerce.string()
})

export const GetServiceResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "slug": zod.string(),
  "description": zod.string(),
  "category": zod.enum(['llm', 'coding', 'automation', 'interface', 'orchestration', 'observability', 'tool', 'memory', 'runtime']),
  "status": zod.enum(['running', 'stopped', 'starting', 'warning']),
  "health": zod.enum(['healthy', 'degraded', 'offline', 'unknown']),
  "version": zod.string(),
  "port": zod.number().int(),
  "image": zod.string(),
  "uptime": zod.string(),
  "cpuPercent": zod.number(),
  "memory": zod.string(),
  "configured": zod.boolean(),
  "connections": zod.number().int(),
  "accent": zod.string()
})


/**
 * @summary Update managed service configuration
 */
export const UpdateServiceParams = zod.object({
  "serviceId": zod.coerce.string()
})

export const updateServiceBodyPortMax = 65535;



export const UpdateServiceBody = zod.object({
  "port": zod.number().int().min(1).max(updateServiceBodyPortMax).optional(),
  "image": zod.string().optional(),
  "configured": zod.boolean().optional()
})

export const UpdateServiceResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "slug": zod.string(),
  "description": zod.string(),
  "category": zod.enum(['llm', 'coding', 'automation', 'interface', 'orchestration', 'observability', 'tool', 'memory', 'runtime']),
  "status": zod.enum(['running', 'stopped', 'starting', 'warning']),
  "health": zod.enum(['healthy', 'degraded', 'offline', 'unknown']),
  "version": zod.string(),
  "port": zod.number().int(),
  "image": zod.string(),
  "uptime": zod.string(),
  "cpuPercent": zod.number(),
  "memory": zod.string(),
  "configured": zod.boolean(),
  "connections": zod.number().int(),
  "accent": zod.string()
})


/**
 * @summary Remove a managed service from the fleet
 */
export const RemoveServiceParams = zod.object({
  "serviceId": zod.coerce.string()
})

export const RemoveServiceResponse = zod.object({
  "success": zod.boolean(),
  "serviceId": zod.string()
})


/**
 * @summary Start, stop, or restart a managed service
 */
export const ControlServiceParams = zod.object({
  "serviceId": zod.coerce.string()
})

export const ControlServiceBody = zod.object({
  "action": zod.enum(['start', 'stop', 'restart'])
})

export const ControlServiceResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "slug": zod.string(),
  "description": zod.string(),
  "category": zod.enum(['llm', 'coding', 'automation', 'interface', 'orchestration', 'observability', 'tool', 'memory', 'runtime']),
  "status": zod.enum(['running', 'stopped', 'starting', 'warning']),
  "health": zod.enum(['healthy', 'degraded', 'offline', 'unknown']),
  "version": zod.string(),
  "port": zod.number().int(),
  "image": zod.string(),
  "uptime": zod.string(),
  "cpuPercent": zod.number(),
  "memory": zod.string(),
  "configured": zod.boolean(),
  "connections": zod.number().int(),
  "accent": zod.string()
})


/**
 * @summary Get recent service logs
 */
export const GetServiceLogsParams = zod.object({
  "serviceId": zod.coerce.string()
})

export const GetServiceLogsResponseItem = zod.object({
  "id": zod.string(),
  "timestamp": zod.string(),
  "stream": zod.enum(['system', 'stdout', 'stderr']),
  "message": zod.string()
})
export const GetServiceLogsResponse = zod.array(GetServiceLogsResponseItem)


/**
 * @summary List service connections
 */
export const ListConnectionsResponseItem = zod.object({
  "id": zod.string(),
  "source": zod.string(),
  "target": zod.string(),
  "protocol": zod.string(),
  "status": zod.enum(['connected', 'pending', 'needs-config']),
  "detail": zod.string()
})
export const ListConnectionsResponse = zod.array(ListConnectionsResponseItem)