# lib/api-zod/src/generated/types/catalogItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogItem {
  id: string;
  name: string;
  tier: string;
  category: string;
  description: string;
  upstreamUrl: string;
  license: string;
  image: string;
  defaultPort: number;
  installed: boolean;
  accent: string;
}