# lib/api-zod/src/generated/types/envVarSpec.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnvVarSpec {
  key: string;
  description: string;
  required: boolean;
  defaultValue?: string;
  secret: boolean;
}