# lib/api-zod/src/generated/types/installServiceRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InstallServiceRequest {
  manifestId: string;
  customPort?: number;
}