# lib/api-zod/src/generated/types/manifestDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EnvVarSpec } from './envVarSpec';
import type { PortMapping } from './portMapping';

export interface ManifestDetail {
  id: string;
  name: string;
  tier: string;
  category: string;
  description: string;
  upstreamUrl: string;
  license: string;
  image: string;
  defaultPort: number;
  ports: PortMapping[];
  env: EnvVarSpec[];
  capabilities: string[];
  compatibleConnections: string[];
  securityNote: string;
}