# lib/api-zod/src/generated/types/portMapping.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PortMapping {
  containerPort: number;
  hostPort: number;
  label: string;
}