# lib/api-zod/src/generated/types/removeServiceResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RemoveServiceResponse {
  success: boolean;
  serviceId: string;
}